analysis_python/requirements.txt
# SilentTrace Python Analysis Layer Dependencies
# Install with: pip install -r requirements.txt

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Data manipulation and analysis
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
plotly>=5.0.0

# Web framework for dashboard
flask>=2.0.0
flask-cors>=3.0.0

# CLI enhancements
colorama>=0.4.4
rich>=10.0.0

# Real-time data handling
websockets>=10.0

# Configuration management
pyyaml>=5.4.0

# Optional: For future ML-based signal classification
# scikit-learn>=1.0.0
# tensorflow>=2.6.0